requirements.txt
# Snowflake FinOps Toolkit - Dependencies
# Updated: 2026
# Install: pip install -r requirements.txt

# Snowflake connector with pandas support
snowflake-connector-python[pandas]==3.12.0

# Streamlit dashboard
streamlit==1.44.0

# Data processing
pandas==2.2.3
numpy==1.26.4

# Visualization
plotly==5.24.1

# Environment variables
python-dotenv==1.0.1

# Key-pair authentication support
cryptography==43.0.3

# Testing
pytest==8.3.3
pytest-cov==6.0.0

# Code quality (optional but recommended)
black==24.10.0
flake8==7.1.1